server/requirements.txt
Flask
python-dotenv
flask-cors
PyJWT
bcrypt
supabase
httpx>=0.24.0
websockets>=13.0
